Добавил спидометр и потимизировал его расположение на странице, убрал скрол страницы
