--- src/container/mining/components/ProgressBar.style.ts
@@ -1,11 +1,0 @@
-import { css } from '@emotion/css';
-
-export const progressBarStyle = css`
-    --red: #fe0222;    
-    top: 0;
-    left: 0;
-    right: 0;
-    height: 10px;
-    background: var(--red);
-    transform-origin: 0%;
-`;
