--- src/container/mining/components/ProgressBar.tsx
@@ -1,17 +1,0 @@
-import  React from 'react';
-import { motion } from "framer-motion";
-import { progressBarStyle } from './ProgressBar.style';
-
-const ProgressBar = ({ progress }): React.ReactElement => {
-    
-    return (
-        <>
-            <motion.div
-            className={progressBarStyle}
-            style={{ scaleX: progress }}
-            />
-        </>
-    );
-};
-
-export default ProgressBar;
